requirements.txt
redis~=4.1.4
apache-airflow
elasticsearch<8.0.0
pendulum~=2.1.2
psycopg2==2.8.6
